+++ test/react/config.babel.js
@@ -1,0 +1,17 @@
+import path from 'path';
+
+module.exports = {
+  entry: [path.join(__dirname, './main.jsx')],
+  module: {
+    loaders: [
+      {
+        test: /\.jsx?$/,
+        exclude: /node_modules/,
+        loader: 'babel',
+        query: {
+          presets: ['es2015', 'react', 'react-hmre']
+        }
+      }
+    ]
+  }
+}